Add DB.Update() method.
We already have a DB.View() method for read-only transactions, so it
makes sense to have a corresponding method for read-write txns.

This method is a convenience wrapper around simple read-write
transactions. We create the transcation, and pass it as an argument to
the func that was passed into the method. We also make sure we
commit/discard the transaction without the caller needing to do it
themselves.