Run gc until we encounter non-nil error.
Earlier we ran the GC only once, relying on the user to call
DB.RunValueLogGC() multiple times.  We now record some discard
statistics to the vlog struct. We can use that information to improve
the GC.

This commit makes the GC run multiple times, picking a different log
file each time, until we encounter an error. We expect to stop the GC
the moment we receive a ErrNoRewrite error.