Bug: Max key size
Fix a bug caused because the max key size enforced by transaction was
the same size enforced by the value log. Due to this, when we add a
badgerMove prefix to the key, it was beyond the uint16 limits, causing
the value log to truncate the file.

Instead of trying to instrument a strict max key size, which might get
broken later, this change sets the size to 65000 bytes, which has around
500 bytes of leeway for various internal alterations. Value log only
enforces that the key be below uint16, with no knowledge of the limit
enforced by txn APIs.

Fixes #583.