Bug Fix: Truncate of penultimate log files (#616)
Value log files minus the most recent log one, were being opened with read-only permissions. If the value head pointer was pointing to them, and one of them had a file system corruption issue, where they missed their bits, it would cause Badger to try to truncate them. Being read-only, this would cause an "invalid argument" to be returned by the Go file.truncate call. This error caused confusion before, but essentially, it was being returned because the files were read-only, and therefore couldn't be truncated (EINVAL).

This PR fixes that by refactoring how value log is opened. It iterates over the log files, and if they need to be replayed, it would open them in RW mode; allowing truncation to happen (this also fixes the issue where if the file is mmapped, it won't be truncated). Then, it would open them in the mode they can be consumed by Badger.

Fixes #613 .