Introduce a NewKeyIterator
In Dgraph, we generally need to iterate over all the versions of a
single key. This PR introduces a NewKeyIterator method, which sets
iterator option's Prefix to the key, and also uses this prefix to do
bloom filter lookups to remove as many tables we can, when creating the
iterator. This should make the iterator.Seek more efficient.