Fix a bug in StreamWriter which causes key overlap in levels. (#824)
Badger currently put the `!badger!head` key in one of the tables at the last level that StreamWriter writes to.  This can cause issues if any other table in that level has an overlapping key range, considering this key did not come from the Stream Framework iteration.

This PR clears out level 1 for this head key. All other keys start from level 2. At the end, this key is written to level 1, so there's no chance of overlap.

Changes:
* Fix a bug in StreamWriter which causes key overlap in levels.
* Francesc's comments. Also, use level 0 for head key.
* Do not introduce the caps ID. Use capital I, small d.