Store discardMap in vlog file when it's too large (#858)
The discardMap stores the discard status for all value log files in the
db and this could get too large when there are many value log files in
badger. When there are many vlog files, the size of discard map could be
more than what we can keep in the LSM tree.

With this commit, we insert the discardMap into badger using the write
channel which would automatically put the discard map into the value log
file and use it's pointer in the LSM tree if the LSM tree cannot hold it.