[breaking/format] Use varint for header encoding (without header length) (#935)
This is a breaking change.

Value log stores an entry for each key-value pair stored in it.
The existing format looks like
+-------------------------------+-----+-------+----------+
| Header (fixed size: 18 bytes) | Key | Value | Checksum |
+-------------------------------+-----+-------+----------+

We don't always use up all the 18 Bytes of the header. This PR proposes
varint encoding for the individual attributes of the header. The varint
encoded header can have length in the range of 5-21 Bytes.

The proposed entry in value log looks like
+------------------------+-----+-------+----------+
| Header (variable size) | Key | Value | Checksum |
+------------------------+-----+-------+----------+