Store entire L0 in memory (#963)
We build L0 from memtables, write it to files and then reopen them. We
can instead build the tables but keep them in memory. This would save the
overhead of writing the tables to the file.

This change has a downside that if the DB crashes, reopening it would take
longer since all the entries in memtable and L0 have to be replayed.

This commit also adds a `KeepL0InMemory` option. When set, all L0 tables
will be in memory. This option also sets the `CompactL0OnClose` flag.
The `KeepL0InMemory` flag is set to true by default.