Introduce StreamDone in Stream Writer (#1061)
This PR introduces, a way to tell StreamWriter to close a stream. Previously Streams were always open until Flush is called on StreamWriter. This resulted in memory utilisation, because of underlying TableBuilder to a sortedWriter. Also closing all sorted writer in single call resulted in 
more memory allocation(during Flush()). This can be useful in some case such as bulk loader in
Dgraph, where only one stream is active at a time.