Introduce in-memory mode in badger (#1113)
This PR introduces in-memory mode in badger. The in-memory mode
can be enabled by setting options.InMemory=true. When badger is
running in in-memory mode no files are created and everything 
is stored in memory.

On DB close, all stored data is lost.
NOTE - An existing DB cannot be opened in in-memory mode.

Fixes - https://github.com/dgraph-io/badger/issues/1001