Update head while replaying value log (#1372)
Fixes https://github.com/dgraph-io/badger/issues/1363

The head pointer is not updated when we perform replays. The head
pointer would be updated only when the replay completes. If badger crashes
between the point when replay started and replay finished, we would end up
replaying all the value log files. This PR fixes this issue.