levels: Cleanup builder resources on building an empty table (#1414)
The compaction process can leak `*table.Builder` instances if the table built
is empty.

This used to be a big problem: before #1409 it would leak out the block
compression/encryption goroutines (that are all blocked on reading from the
`b.blockChan`, and because of that, it would leak the builder too.

Since #1409 this is now a simple question of correctness. Take it or leave it.