feat(externalMagic): Introduce external magic number (#1745)
Badger writes a magic number in the manifest. It helps badger to decide if its current version is compatible with the data on the disk or not. But now the internal data storage of badger is not changed but the data format for dgraph has changed. This causes a problem if someone starts dgraph-2109 on an older dgraph directory. This change adds an external magic number to badger which will help avoid data corruption by causing panic if opening Dgraph on the wrong directory.

There are 8 magic bytes in the manifest.
Prior to this change:

0-4: magic text
4-8: badgerMagicNumber
After this change:

0-4: magic text
4-6: externalMagicNumber
6-8: badgerMagicNumber